Refactor: 디렉토리 이름 수정 * 변경 사항: router -> routes * 변경 이유: 디렉토리간 이름 단/복수 맞추기 위해
